refactor(scripts): split gen-llms catalog builders by domain
Extract errors/advisories builders into llms-diagnostic-docs.ts and the
lifecycle guide builder into llms-lifecycle-docs.ts. gen-llms keeps
guidePages, llms surfaces, and the public re-export surface. Register
the new modules on the docs CI lane and content-hash inputs.
