+++ scripts/gen-llms.ts
@@ -6,22 +6,18 @@
  * Module layout:
  * - `llms-markdown.ts` — minimal markdown renderer
  * - `llms-guide-content.ts` — guide prose constants + interaction reference
- * - this file — catalog-driven builders, guidePages, llms surfaces, re-exports
+ * - `llms-diagnostic-docs.ts` — errors + advisories catalog builders
+ * - `llms-lifecycle-docs.ts` — lifecycle guide builder + LifecycleDoc
+ * - this file — guidePages, llms surfaces, public re-exports
  *
  * Everything here is pure (data in, string out) and unit-tested in
  * scripts/gen-llms.test.ts; apps/docs imports it via the `$scripts` alias
  * for its prerendered endpoints and guide pages.
  */
-import { ADVISORY_CATALOG } from "@ggsvelte/core";
-import { CURRENT_EDITION, LINT_CATALOG, THEME_NAMES } from "@ggsvelte/spec";
+import { CURRENT_EDITION, THEME_NAMES } from "@ggsvelte/spec";
 import sveltePackage from "../packages/svelte/package.json";
 import { GUIDE_CATALOG, type GuideSlug } from "../apps/docs/src/lib/catalog/guide";
 import { interactionExpositionSlug } from "../apps/docs/src/lib/catalog/interaction-exposition";
-import {
-  buildDiagnosticDocs,
-  type DiagnosticDocEntry,
-  type DiagnosticDocSource,
-} from "./diagnostic-docs";
 import { assertGuideCodeContract } from "./guide-code-contract";
 import {
   FACETS_COORDINATES_MD,
@@ -34,6 +30,8 @@ import {
   TEMPORAL_SCALES_MD,
   UPGRADING_MD,
 } from "./llms-guide-content";
+import { buildAdvisoriesMd, buildErrorsMd } from "./llms-diagnostic-docs";
+import { buildLifecycleMd, type LifecycleDoc } from "./llms-lifecycle-docs";
 
 export { buildDiagnosticDocs } from "./diagnostic-docs";
 export { extractMarkdownHeadings, renderMarkdown, type MarkdownHeading } from "./llms-markdown";
@@ -50,208 +48,9 @@ export {
   UPGRADING_MD,
   type InteractionReferenceEntry,
 } from "./llms-guide-content";
+export { buildAdvisoriesMd, buildErrorsMd } from "./llms-diagnostic-docs";
+export { buildLifecycleMd, type LifecycleDoc } from "./llms-lifecycle-docs";
 
-function catalogSection(
-  title: string,
-  intro: string,
-  catalog: Record<string, { summary: string; fix?: string }>,
-  opts: { tierOf?: (code: string) => string } = {},
-): string {
-  const lines = [`## ${title}`, "", intro, ""];
-  for (const [code, entry] of Object.entries(catalog)) {
-    const tier = opts.tierOf === undefined ? "" : ` (${opts.tierOf(code)})`;
-    lines.push(`### \`${code}\`${tier}`, "", entry.summary, "");
-    if (entry.fix !== undefined) lines.push(`**Fix:** ${entry.fix}`, "");
-  }
-  return lines.join("\n");
-}
-
-const diagnosticSectionTitles: Record<DiagnosticDocSource, string> = {
-  validation: "Validation errors (@ggsvelte/spec)",
-  pipeline: "Render-time errors (@ggsvelte/core)",
-  warning: "Render warnings",
-  interaction: "Interaction diagnostics (@ggsvelte/svelte)",
-  cli: "CLI diagnostics (ggsvelte-render)",
-};
-
-function diagnosticHeading(entry: DiagnosticDocEntry): string {
-  const primaryAnchor = entry.code
-    .toLowerCase()
-    .replaceAll(/[^a-z0-9]+/g, "-")
-    .replaceAll(/^-|-$/g, "");
-  return entry.anchor === primaryAnchor
-    ? `### \`${entry.code}\``
-    : `### \`${entry.code}\` — ${entry.source}`;
-}
-
-function renderDiagnosticEntry(entry: DiagnosticDocEntry): string {
-  const lines = [
-    diagnosticHeading(entry),
-    "",
-    `**Code + severity:** \`${entry.code}\` · ${entry.severity}`,
-    "",
-    `**What failed:** ${entry.whatFailed}`,
-    "",
-    `**Why:** ${entry.why}`,
-    "",
-    `**Fix:** ${entry.fix}`,
-    "",
-    `**Consequence (${entry.consequence}):** ${entry.consequenceText}`,
-    "",
-    `**Stable link:** [/guide/errors#${entry.anchor}](/guide/errors#${entry.anchor})`,
-    "",
-  ];
-  if (entry.recipe !== undefined) {
-    lines.push(
-      "**Minimal illustration — copy only the relevant fragment:**",
-      "",
-      `\`\`\`${entry.recipe.language} fragment copy`,
-      entry.recipe.code,
-      "```",
-      "",
-    );
-  }
-  return lines.join("\n");
-}
-
-export function buildErrorsMd(): string {
-  const entries = buildDiagnosticDocs();
-  const sections = (Object.keys(diagnosticSectionTitles) as DiagnosticDocSource[]).map((source) => {
-    const intro =
-      source === "cli"
-        ? "SVG is written only to stdout; JSON Lines diagnostics are written only to stderr. Exit 1 means rendering failed, exit 2 means usage/input failed, and exit 3 means spec validation failed."
-        : "Each entry answers what failed, why, how to recover safely, and whether output was blocked or degraded.";
-    return [
-      `## ${diagnosticSectionTitles[source]}`,
-      "",
-      intro,
-      "",
-      ...entries
-        .filter((entry) => entry.source === source)
-        .map((entry) => renderDiagnosticEntry(entry)),
-    ].join("\n");
-  });
-
-  return `# Errors reference
-
-Diagnostics are generated from the catalogs used by validation, rendering,
-interaction, and the CLI. Identity is the pair \`(source, code)\`: a bare code
-can intentionally exist in more than one source with a different consequence.
-
-## Quickstart troubleshooting
-
-- **Collapsed or zero-width container:** the responsive plot remains
-  \`data-gg-ready="false"\` until ResizeObserver reports a positive width.
-  Give the parent a real grid/flex track width; no fixed chart width is needed.
-- **SSR and hydration:** omitted width server-renders at 832 × 400, stays
-  not-ready on the server, then measures its real container after hydration.
-- **Unexpected height:** omitted height is 400px unless the spec supplies one.
-- **TypeScript or linked-package mismatch:** install one compatible
-  \`@ggsvelte/svelte\` version and let it resolve matching core/spec packages;
-  remove stale lockfile overrides that mix versions.
-- **CLI input failure:** run \`ggsvelte-render --help\`; keep SVG stdout
-  separate from JSON Lines stderr while correcting the reported input.
-
-${sections.join("\n\n")}
-`;
-}
-
-export function buildAdvisoriesMd(): string {
-  const lint = catalogSection(
-    "Spec-lint advisories (@ggsvelte/spec lintSpec)",
-    'Valid-but-questionable specs (Hadley: "we can produce many plots that don\'t make sense, yet are grammatically valid"). Run `lintSpec(spec, { profile? })` directly, pass `{ lint: true }` to `validate()`, or read the CLI\'s stderr advisories (source "spec-lint"). Data-dependent rules skip silently without evidence.',
-    LINT_CATALOG,
-    {
-      tierOf: (code) => `needs: ${LINT_CATALOG[code as keyof typeof LINT_CATALOG].needs}`,
-    },
-  );
-  const heuristics = catalogSection(
-    "Pipeline heuristic advisories (@ggsvelte/core)",
-    "Every heuristic decision the pipeline takes is disclosed as `{ code, path, chosen, howToOverride }` on `RenderModel.advisories` — agents see the guess and can correct it.",
-    ADVISORY_CATALOG,
-  );
-  return `# Advisories
-
-Advisories never block a render. Two distinct kinds, two sources:
-
-${lint}
-
-${heuristics}
-`;
-}
-
-/** Shape of the generated lifecycle.json document (scripts/gen-lifecycle.ts). */
-export interface LifecycleDoc {
-  tags: readonly string[];
-  surfaces: readonly {
-    package: string;
-    entry: string;
-    source: string;
-    exports: Record<string, { kind: "value" | "type"; lifecycle: string }>;
-  }[];
-}
-
-export function buildLifecycleMd(lifecycle: LifecycleDoc): string {
-  const sections = lifecycle.surfaces.map((surface) => {
-    const byTag = new Map<string, string[]>();
-    for (const [name, meta] of Object.entries(surface.exports)) {
-      const list = byTag.get(meta.lifecycle) ?? [];
-      list.push(meta.kind === "type" ? `\`${name}\` (type)` : `\`${name}\``);
-      byTag.set(meta.lifecycle, list);
-    }
-    const parts = [
-      `## ${surface.package}${surface.entry === "." ? "" : ` (${surface.entry})`}`,
-      "",
-    ];
-    for (const tag of lifecycle.tags) {
-      const names = byTag.get(tag);
-      if (names === undefined) continue;
-      parts.push(`### ${tag} (${String(names.length)})`, "", names.join(", "), "");
-    }
-    return parts.join("\n");
-  });
-  return `# Lifecycle & editions
-
-## Lifecycle tags
-
-Every public export carries a lifecycle tag (generated into
-\`lifecycle.json\` from the source annotations — regenerate with
-\`bun run lifecycle:gen\`):
-
-- **experimental** — may change or disappear in any 0.x release. The default
-  for APIs not explicitly promoted.
-- **stable-intent** — committed enough that changes pay the migration tax.
-  Covers (1) the agent core path (PortableSpec, normalize, validate,
-  renderToSVGString, GGPlot and their direct result contracts) and
-  (2) the recommended Svelte composition children (\`Geom*\` shells and the
-  theme/scale/coord/facet/labs/guides/legend grammar children). Not frozen
-  pre-1.0, but changes here are treated as breaking: they get a changeset, a
-  migration note, and a deprecation window where feasible. Registry and
-  factory helpers stay experimental.
-- **stable** — committed API under semver (none in v0.1).
-- **superseded** — keeps working but stops being recommended; docs point to
-  the replacement. Protects agent-generated code from silent breakage.
-
-## Defaults editions
-
-\`normalize()\` stamps \`edition: ${String(CURRENT_EDITION)}\` onto every spec that doesn't carry one,
-freezing which generation of DEFAULT aesthetics (theme role tokens,
-categorical palette, sequential ramp) the spec was authored against. When a
-future edition ships better defaults, stamped specs keep their original look
-— old charts never reshuffle. Explicit settings (\`theme\`,
-\`scales.*.range\`, \`scales.*.scheme\`) always win over edition defaults, and
-unknown editions degrade to the latest known with an \`unknown-edition\`
-warning.
-
-${sections.join("\n")}
-`;
-}
-
-// ---------------------------------------------------------------------------
-// llms.txt / llms-full.txt
-// ---------------------------------------------------------------------------
-
-/** Manifest-entry shape consumed here (structural subset of the manifest). */
 export interface LlmsExampleEntry {
   id: string;
   title: string;
